series: move rewrites for Abs/Piecewise from the gruntz.py
